move ValidateBase to base file
